+++ tests/Model/Data3.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of rekalogika/temporary-url-bundle package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\TemporaryUrl\Tests\Model;
+
+class Data3
+{
+    public function __construct(
+        private string $content,
+    ) {
+    }
+
+    public function getContent(): string
+    {
+        return $this->content;
+    }
+}