+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.1
+
+* Use regex for Twig testing
+* Add tests for Union types
+
 ## 1.1.0
 
 * Add 'temporary_url' Twig filter